Created Script for moving platforms
Completely customizable moving platforms. You can set the speed they move, the range and even the direction they start moving in.
NOTE: These platforms are incapable of starting in the middle so keep that in mind
and they designed to move left and right XOR up and down do not turn on both at the same time.